requirements.txt
# Deep learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Experiment tracking
wandb>=0.16.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data & metrics
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Config & utilities
pyyaml>=6.0
tqdm>=4.65.0

# Model hosting
huggingface-hub>=0.19.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
